basic functions and order begin

+++ src/database/migrations/20230619182507-create-product.js
@@ -14,7 +14,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.REAL,
         allowNull: false,
       },
       brand: {
